backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
sqlalchemy==2.0.25
psycopg2-binary>=2.9.9
openpyxl==3.1.2
pandas==2.2.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.115.0
aiofiles==23.2.1
python-dotenv==1.0.1
email-validator>=2.0
